+++ Right Posture/modules/__init__.py
@@ -51,4 +51,4 @@ from .app_auth import Auth_system
 from .app_charts import Line_charts
 
 # APP SETTING
-from .app_setting_page import Main_data, save_checkbox
+from .app_data import Main_data, save_checkbox
